refactor serve app
